requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
dj-database-url==3.0.0
Django==4.2.23
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.0
whitenoise==6.9.0
